rework custom jekyll plugin into rake task
This rake task will update the _config.yml with fresh stats,
commit afterwards and push back to repo.

Currently everything except the push is implemented, including some
debug output.
