+++ Gemfile
@@ -11,3 +11,4 @@ gem 'html-proofer'
 gem 'rake'
 gem 'rubocop'
 gem 'rubocop-rspec'
+gem 'safe_yaml'
